chore(core): move releases tool to standalone plugin
