+++ packages/sanity/src/core/scheduled-publishing/tool/Tool.tsx
@@ -8,10 +8,10 @@ import {LoadingBlock} from '../../components/loadingBlock/LoadingBlock'
 import {TimeZoneButton} from '../../components/timeZone/timeZoneButton/TimeZoneButton'
 import {useTimeZone} from '../../hooks/useTimeZone'
 import {useTranslation} from '../../i18n/hooks/useTranslation'
-import {useReleasesToolAvailable} from '../../releases/hooks/useReleasesToolAvailable'
 import {useScheduledDraftsEnabled} from '../../releases/hooks/useScheduledDraftsEnabled'
 import {RELEASES_SCHEDULED_DRAFTS_INTENT} from '../../releases/plugin'
 import {useScheduledPublishingEnabled} from '../../scheduledPublishing/contexts/ScheduledPublishingEnabledProvider'
+import {useReleasesToolAvailable} from '../../schedules/hooks/useReleasesToolAvailable'
 import {useWorkspace} from '../../studio/workspace'
 import TimeZoneButtonElementQuery from '../components/dialogs/TimeZoneButtonElementQuery'
 import ErrorCallout from '../components/errorCallout/ErrorCallout'
